Add SQL functions and user defined types to work with filesystem objects: files and directories. Added new and refactored existing methods to work with streams.

+++ PowerLib.System/IO/Streamed/Typed/NulRangeStreamedRegularArray.cs
@@ -87,8 +87,8 @@ namespace PowerLib.System.IO.Streamed.Typed
     {
       if (value.HasValue)
       {
-        stream.WriteDouble(value.Value.Index);
-        stream.WriteDouble(value.Value.Count);
+        stream.WriteInt32(value.Value.Index);
+        stream.WriteInt32(value.Value.Count);
       }
     }
